+++ src/Tags/TagInterface.php
@@ -1,0 +1,11 @@
+<?php declare(strict_types=1);
+
+namespace Becklyn\RadBundle\Tags;
+
+interface TagInterface
+{
+    /**
+     * Returns the tag label.
+     */
+    public function getLabel () : string;
+}